+++ e2e/commands/build.e2e.1.js
@@ -34,17 +34,24 @@ describe('bit build', function () {
       const output = helper.command.build();
       expect(output).to.have.string('nothing to build');
     });
-    describe('after importing a compiler', () => {
+    describe('build after importing a compiler', () => {
+      let buildOutput;
       before(() => {
         const output = helper.env.importCompiler();
         expect(output).to.have.string(
           `the following component environments were installed\n- ${helper.scopes.env}/compilers/babel@`
         );
+        buildOutput = helper.command.build();
       });
       it('should successfully import and build using the babel compiler', () => {
-        const buildOutput = helper.command.build();
         expect(buildOutput).to.have.string(path.normalize('dist/bar/foo.js.map'));
         expect(buildOutput).to.have.string(path.normalize('dist/bar/foo.js'));
+      });
+      it('should create links to node_modules with partial-ids (id without scope)', () => {
+        const nodeModuleDir = path.join(helper.scopes.localPath, 'node_modules/@bit/bar.foo');
+        expect(nodeModuleDir).to.be.a.directory();
+        expect(path.join(nodeModuleDir, 'package.json')).to.be.a.file();
+        expect(path.join(nodeModuleDir, 'bar/foo.js')).to.be.a.file();
       });
       describe('when an exception is thrown during the build', () => {
         before(() => {
@@ -125,6 +132,22 @@ describe('bit build', function () {
             expect(distFileFullPath).to.be.a.file();
             expect(compilerFolderFullPath).to.be.a.directory().and.not.empty;
           });
+        });
+      });
+      describe('after exporting the component', () => {
+        before(() => {
+          helper.command.tagAllComponents();
+          helper.command.exportAllComponents();
+        });
+        it('should delete the generated links on node-modules', () => {
+          const nodeModuleDir = path.join(helper.scopes.localPath, 'node_modules/@bit/bar.foo');
+          expect(nodeModuleDir).to.not.be.a.path();
+        });
+        it('should generate new links on node-modules with full-id (include scope name)', () => {
+          const nodeModuleDir = path.join(helper.scopes.localPath, `node_modules/@bit/${helper.scopes.remote}.bar.foo`);
+          expect(nodeModuleDir).to.be.a.directory();
+          expect(path.join(nodeModuleDir, 'package.json')).to.be.a.file();
+          expect(path.join(nodeModuleDir, 'bar/foo.js')).to.be.a.file();
         });
       });
     });